// src/api/user/output/get-self-data-output.ts
export type GetSelfDataOutput = {
  id: string;
  email: string;
  role: string;
  createdAt: string;
  updatedAt: string;
  contact: {
    id: string;
    name: string;
    surname: string;
    phone: string;
    cpf: string;
    bornDate: string;
    userId: string;
    createdAt: string;
    updatedAt: string;
  };
};